+++ src/library/DIPS.Mobile.UI/Components/Labels/Label.cs
@@ -1,13 +1,11 @@
-using DIPS.Mobile.UI.Extensions;
-using DIPS.Mobile.UI.Resources.Colors;
-
 namespace DIPS.Mobile.UI.Components.Labels
 {
-    public class Label : Microsoft.Maui.Controls.Label
+    public partial class Label : Microsoft.Maui.Controls.Label
     {
         public Label()
         {
             this.SetAppThemeColor(TextColorProperty, ColorName.color_system_black);
+            MaxLines = int.MaxValue;
         }
     }
 }